requirements.txt
bentoml==1.0.10
joblib==1.2.0
pip-chill==1.0.1
torchvision==0.22.1
tqdm==4.64.1
wincertstore==0.2
torchsummary==1.5.1
#matplotlib==3.5.1
opencv-python
fastapi
uvicorn